Add a nested route for categories under an application.
There are some problems here, namely with the social auth router, which now does not work. Further, the categories need to have their own dedicated endpoint because cannot seem to make it read-write once they're nested under an application. Ultimately, I'm not convinced we want to keep this functionality and may be better off using layered url patterns and filters instead but the concept is functioning here so commiting it.