Add save validation to when selecting default credentials for a cloud.
Via Admin, this works to ensure one and only one default set of credentials is selected for a cloud. However, it does not work across all clouds. For example, a user can add a default set of credentials for cloud A but not select a default for Cloud B. Enforcing this at the formset is probably doable but that will not affect the API so opting not to worry about it for the time being.