Update network selection during the launch process
These changes account for some changes in CloudBridge and how it
handles networks. In the past, CloudBridge required a subnet ID for
AWS and a network ID for OpenStack. This was changed to always
require the network ID. However, this came at the expense of not
being able to select a subnet (hence, we should just remove that
field from the list of choices).

Also, this requires CloudBridge at rev (or later)
https://github.com/gvlproject/cloudbridge/commit/1e2e85cd5df1bab732290b3b617c520b5cf4abc5